Use java logger in demo App, remove unused classes
