+++ app/src/main/java/io/horizontalsystems/ethereumkit/sample/App.kt
@@ -1,18 +1,20 @@
 package io.horizontalsystems.ethereumkit.sample
 
 import android.app.Application
-import android.util.Log
 import com.facebook.stetho.Stetho
 import io.reactivex.plugins.RxJavaPlugins
+import java.util.logging.Logger
 
 class App : Application() {
+
+    private val logger = Logger.getLogger("App")
 
     override fun onCreate() {
         super.onCreate()
         instance = this
 
         RxJavaPlugins.setErrorHandler { e: Throwable? ->
-            Log.w("RxJava ErrorHandler", e)
+            logger.warning("RxJava ErrorHandler: ${e?.message}")
         }
 
         // Enable debug bridge
