+++ erc20kit/src/main/java/io/horizontalsystems/erc20kit/models/Erc20LogEvent.kt
