+++ ethereumkit/src/main/java/io/horizontalsystems/ethereumkit/core/TransactionManager.kt
@@ -1,116 +1,59 @@
 package io.horizontalsystems.ethereumkit.core
 
-/*class TransactionManager(
-        private val storage: ITransactionStorage,
-        private val transactionsProvider: ITransactionsProvider
-) : ITransactionManager {
+import io.horizontalsystems.ethereumkit.models.Address
+import io.horizontalsystems.ethereumkit.models.FullTransaction
+import io.horizontalsystems.ethereumkit.models.Transaction
+import io.horizontalsystems.ethereumkit.transactionsyncers.TransactionSyncManager
+import io.reactivex.BackpressureStrategy
+import io.reactivex.Flowable
+import io.reactivex.Single
+import io.reactivex.disposables.CompositeDisposable
+import io.reactivex.schedulers.Schedulers
+import io.reactivex.subjects.PublishSubject
 
-    private var delayTime = 3 // in seconds
-    private val delayTimeIncreaseFactor = 2
-    private var retryCount = 0
+class TransactionManager(
+        private val address: Address,
+        private val transactionSyncManager: TransactionSyncManager,
+        private val storage: ITransactionStorage
+) {
+    private val disposables = CompositeDisposable()
+    private val etherTransactionSubject = PublishSubject.create<List<FullTransaction>>()
+    private val allTransactionsSubject = PublishSubject.create<List<FullTransaction>>()
 
-    private var disposables = CompositeDisposable()
-
-    override var syncState: SyncState = SyncState.NotSynced(SyncError.NotStarted())
-        private set(value) {
-            if (field != value) {
-                field = value
-                listener?.onUpdateTransactionsSyncState(value)
-            }
-        }
-
-    override val source: String
-        get() = transactionsProvider.source
-
-    override var listener: ITransactionManagerListener? = null
-
-    private fun update(transactions: List<EtherscanTransaction>, internalTransactions: List<InternalTransaction>) {
-        storage.saveTransactions(transactions)
-        storage.saveInternalTransactions(internalTransactions)
-
-        val transactionsWithInternal = transactions.mapNotNull { transaction ->
-            val internalTransactionList = internalTransactions.filter { it.hash.contentEquals(transaction.hash) }
-            val transactionWithInternal = TransactionWithInternal(transaction, internalTransactionList)
-
-            if (!isEmpty(transactionWithInternal)) {
-                transactionWithInternal
-            } else {
-                null
-            }
-        }
-
-        listener?.onUpdateTransactions(transactionsWithInternal)
-    }
-
-    private fun isEmpty(transactionWithInternal: TransactionWithInternal): Boolean {
-        return transactionWithInternal.transaction.value == BigInteger.ZERO && transactionWithInternal.internalTransactions.isEmpty()
-    }
-
-    private fun sync(delayTimeInSeconds: Int? = null) {
-        val lastTransactionBlockHeight = storage.getLastTransactionBlockHeight() ?: 0
-        val lastInternalTransactionBlockHeight = storage.getLastInternalTransactionBlockHeight() ?: 0
-
-        var requestsSingle = Single.zip(
-                Single.just(listOf()),
-                Single.just(listOf()),
-//                transactionsProvider.getTransactions(lastTransactionBlockHeight + 1),
-//                transactionsProvider.getInternalTransactions(lastInternalTransactionBlockHeight + 1),
-                BiFunction<List<EtherscanTransaction>, List<InternalTransaction>, Pair<List<EtherscanTransaction>, List<InternalTransaction>>> { t1, t2 -> Pair(t1, t2) }
-        )
-
-        delayTimeInSeconds?.let {
-            requestsSingle = requestsSingle.delaySubscription(delayTimeInSeconds.toLong(), TimeUnit.SECONDS)
-        }
-
-        requestsSingle
+    init {
+        transactionSyncManager.transactionsAsync
                 .subscribeOn(Schedulers.io())
-                .subscribe({ (transactions, internalTransactions) ->
-                    if (retryCount > 0 && transactions.isEmpty() && internalTransactions.isEmpty()) {
-                        retryCount -= 1
-                        delayTime *= delayTimeIncreaseFactor
-                        sync(delayTime)
-                        return@subscribe
-                    }
-                    retryCount = 0
-                    update(transactions, internalTransactions)
-                    syncState = SyncState.Synced()
-                }, { error ->
-                    retryCount = 0
-                    syncState = SyncState.NotSynced(error)
-                })
-                .let {
-                    disposables.add(it)
+                .subscribe { transactions ->
+                    etherTransactionSubject.onNext(transactions.filter { it.hasEtherTransfer(address) })
+                    allTransactionsSubject.onNext(transactions)
                 }
+                .let { disposables.add(it) }
     }
 
-    override fun refresh(delay: Boolean) {
-        if (syncState is SyncState.Syncing) {
-            return
+    val etherTransactionsAsync: Flowable<List<FullTransaction>> = etherTransactionSubject.toFlowable(BackpressureStrategy.BUFFER)
+    val allTransactionsAsync: Flowable<List<FullTransaction>> = allTransactionsSubject.toFlowable(BackpressureStrategy.BUFFER)
+
+    fun getEtherTransactionsAsync(fromHash: ByteArray? = null, limit: Int? = null): Single<List<FullTransaction>> {
+        return storage.getEtherTransactionsAsync(address, fromHash, limit)
+    }
+
+    fun handle(transaction: Transaction) {
+        storage.save(transaction)
+
+        val fullTransaction = FullTransaction(transaction)
+        if (fullTransaction.hasEtherTransfer(address)) {
+            etherTransactionSubject.onNext(listOf(fullTransaction))
         }
-        syncState = SyncState.Syncing()
+        allTransactionsSubject.onNext(listOf(fullTransaction))
 
-        if (delay) {
-            retryCount = 5
-            delayTime = 3
-
-            sync(delayTime)
-        } else {
-            sync()
-        }
     }
 
-    override fun getTransactions(fromHash: ByteArray?, limit: Int?): Single<List<TransactionWithInternal>> {
-        return storage.getTransactions(fromHash, limit)
+    fun getFullTransactions(fromHash: ByteArray?): List<FullTransaction> {
+        return storage.getFullTransactionsAfter(fromHash)
     }
 
-    override fun handle(transaction: EtherscanTransaction) {
-        storage.saveTransactions(listOf(transaction))
-
-        val transactionWithInternal = TransactionWithInternal(transaction)
-
-        if (!isEmpty(transactionWithInternal)) {
-            listener?.onUpdateTransactions(listOf(transactionWithInternal))
-        }
+    fun getFullTransactions(hashes: List<ByteArray>): List<FullTransaction> {
+        return storage.getFullTransactions(hashes)
     }
 
-}*/
+}
