--- ethereumkit/src/main/java/io/horizontalsystems/ethereumkit/models/TransactionStatus.kt
@@ -1,9 +1,0 @@
-package io.horizontalsystems.ethereumkit.models
-
-enum class TransactionStatus(val value:String) {
-
-    SUCCESS("Success"),
-    FAILED("Failed"),
-    PENDING("Pending"),
-    NOTFOUND("Not Found")
-}
