--- ethereumkit/src/main/java/io/horizontalsystems/ethereumkit/models/TransactionWithInternal.kt
@@ -1,11 +1,0 @@
-package io.horizontalsystems.ethereumkit.models
-
-import androidx.room.Embedded
-import androidx.room.Relation
-
-data class TransactionWithInternal(
-        @Embedded
-        val transaction: EtherscanTransaction,
-        @Relation(parentColumn = "hash", entityColumn = "hash")
-        val internalTransactions: List<InternalTransaction> = listOf()
-)
